Added backing collection for `StaticTypeLookupService`
Previously, the `StaticTypeLookupService` inherited from `KeyedCollection<string, Type>`, thus making it act like a collection. This was a convenient shorthand since it needs a collection of types. But it violates best practices since it acts like a collection but isn't named e.g. `StaticTypeLookupCollection`. This correctly raised warning `CA1710`. To remedy this, a new `TypeCollection` was created, and added as a backing field on the `StaticTypeLookupService`. Then, a (`protected`) `Add()` and `Contains()` method were added so derived types could access that collection. And, finally, the existing (`internal`) `TypeCollection` was renamed to the more accurate identiier `TypeMemberInfoCollection` to avoid confusion (it is, in fact, a collection of `MemberInfoCollection` objects keyed by `Type`—and not actually a collection of `Type` instances, as the original name would have suggested.)