Renamed `Error` action to `Index` action
This was already done with the *ASP.NET MVC Framework* version, but hadn't been applied to the *ASP.NET Core MVC* version. This makes the default error route `/Error` instead of `/Error/Error`, thus taking advantage of the default convention of `Index` representing the default action—a convention that's baked into the default route.