Introduced `Remove()` method
We may want to revisit this design later. While we removed the foundation of `KeyedCollection<string, Type>`, we've reintroduced a lot of collection-like syntax (`Add()`, `Contains()`, `Remove()`). There might be a better way of addressing this. That said, this maintains the features of the previous interface that we are using, while reducing the surface area of the API considerably.