Removed obsolete methods
Initially, the `Contract` class sought to maintain interface compatibility with all methods used by the legacy *Code Contracts* implementation as a temporary stop gap allowing the library to be removed. Now that *Code Contracts* has been removed, and all references to deprecated features have been removed or rewritten, those obsolete methods can also be removed. (Since this is an internal class, there is no risk of external dependencies; we can be assured that removing these won't introduce downstream problems at this point.) Methods that continue to be used, such as `Requires<>()`, are maintained—though, eventually, these should be moved to a new class with a more appropriate name to avoid confusion with the legacy *Code Contracts* implementation (e.g., `ParameterValidation`, or something like that).