Resolved bug in reading blob attributes
A recent commit, bf51bd2, introduced an `XmlReader` in preference over the previous `XmlDocument`. While faster, and safer, this also introduced a new bug because both `ReadToFollowing()` _and_ `ReadInnerXml()` advance the position of the reader. As a result, only every other attribute would be read. This is resolved by, instead, using a `do {} while (reader.Name == "attribute")` which prevents the loop itself from advancing the reader. This resolves issues with some blob attributes not showing up, which was causing problems in both the public interface as well as the editor.