Introduced the `[MapToType()]` attribute
The `[MapToType()]` attribute specifies the target type to map the topics to. By default, the type is determined based on the topic's content type. With this attribute, it will either be implicitly set to the type of the property, or will be explicitly set to a type provider. This only represents the attribute; the business logic is not yet built into the `TopicMappingProvider`.