Excluded hidden topics from maping
Hidden topics should not be returned to the user interface. This includes, for example, `List`s, which represent Nested Topics. In those cases, the content of the `List` should be returned as part of any collections mapping to the Nested Set, but the `List` container itself should never be returned (e.g., as part of a `Children` collection). Even beyond the `List` container, any topic explicitly marked as `IsHidden` should have a reasonable expectation of never being returned via a view model.