Established initial interface definition
This definition may need to evolve as the design of the concrete implementation is fleshed out, and new requirements emerge. For now, however, this covers the basic aspects that we'd expect a reverse mapping service to cover. The big unknown right now is how to handle relationships, since that gets into the merging question.