Introduce `ITopicBindingModel`
In the initial version of the interface, each `MapAsync()` method required an `ITopicViewModel`. That was a decent place to start, but there are two issues. First, not all of the properties on `ITopicViewModel` will be required by `IReverseTopicMappingService` implementations. And, second, it is not a best practice for view models to be repurposed as binding models unless all properties are expected to be reused—and, in the case of topics, this is definitely not the case. As such, to avoid accidental or intentional mapping of `*TopicViewModel` classes, the `IReverseTopicMappingService` will utilize its own interface.