Moved `TopicMappingService` to temporary location
In order to maintain the file history of `TopicMappingService` in _both_ `TopicMappingService` _and_ `ReverseTopicMappingService`, we need to revert to the original branch (previous commit) and then temporarily move the original file to a new location (this commit). This is an interstitial step—and, unfortunately, a bit ugly in terms of the git commit history.