Updated interface to `IReverseTopicMappingService` [BREAKING]
This is an intersitial update that _will not compile_! This forces the public (and, actually, private) members to adhere to the new `IReverseTopicMappingService` interface—and makes comparable changes to the private helper functions that they depend upon. This includes renaming `Topic source` to `ITopicBindingModel bindingModel`, `object target` to `Topic topic`, removed `ConcurrentDictionary<int, object> cache` (and the `private` methods that relayed them), and made the most _basic_ updated to the XML doc comments (e.g., to the `<param/>` and `<paramref />` elements) and variables (e.g., renamed any references to `topic` to `source` for clarity.

THIS CODE WILL NOT COMPILE. IT IS EXPECTED TO BE FOLLOWED BY A SERIES OF UPDATES THAT WILL RESOLVE THESE ISSUES.