Removed bypass for mapping of `Topic`
With the `ReverseTopicMappingService` the target is _always_ a `Topic` and it isn't possible to pass in a `Topic` as a source since the source is now an `ITopicBindingModel` and not an `object`. As a result, the bypass from `TopicMappingService` which allowed a `Topic` to be sent directly to a view model that might have a `Topic` property no longer makes much sense and can be safely removed.