Flipped source of attributes
In the `TopicMappingService`, the target view model acted as the source for properties, thus _pulling_ attributes _from_ the source `Topic`. In the `ReverseTopicMappingService`, the model _still_ acts as the source for attributes—but the model is now the _source_. This requires additional validation (which will be added later), but "flipping the script" on this is an easy start to get us on the right track.