Updated reference handling to work with topics
In `TopicMappingService`, a reference `Topic` was looked up by `topic.Id`, mapped to a view model, and set to the property (using `SetTopicReferenceAsync()`.  With the `ReverseTopicMappingService` this is a bit simpler. A reference is simply any object that implements the new `IRelatedTopicBindingModel` interface, and thus exposes a `UniqueKey`. The target `Topic` is looked up using that `UniqueKey` (very similar), but then it's simply set to the `topic.Attributes` collection using `SetInteger`. In practice, these types of attribute references always end in `Id`, even though the Binding Model may not follow that convention. The current code assumes an alias will be setup using `[AttributeKey()]`. This may be updated in the future to enforce this convention, even if the alias isn't setup.