Remove support for `[Flatten]`
The `[Flatten]` attribute allows a hierarchy of child objects to be rolled up into a single index collection. This can act as a predefined "search" engine by filtering all child topics by particular properties. This is really useful for produced views. It doesn't make sense for a Binding Model, however, where the relationship integrity is important, not optimizing the model for the displayed content. As a result, this is being removed.