Removed support for `Get{Property}()`
`TopicMappingService` includes a feature which permits a value to be retrieved from a method named after the attribute. This is useful for converting the value to a more view-friendly name. With a Binding Model, however, we don't expect conversion—we expect to have the data submitted in the format expected. This may be reintroduced should new requirements emerge. But, for now, it's extra complexity that _shouldn't_ be needed.