Removed legacy validation
In `TopicMappingService`, when mapping to a view model, it checks to confirm the target property is settable using the `TypeMemberInfoCollection`. In the `IReverseTopicMappingService` scenario this isn't necessary, as we can always set the value via `topic.Attributes.SetAttribute()`. That said, we _should_ still validate that the requested attribute is _valid_ according to the corresponding `ContentTypeDescriptor`; that work is yet to be done.