Update default value logic
In `TopicMappingService`, the `[DefaultValue()]` attribute can be applied to most any property, and that was done before other attributes were set. It can still be set to most any property, but it really only makes sense on scalar values that are targeting the `topic.Attributes` collection—and only then if there isn't a value otherwise available. As such, this is moved to the `SetScalarValue()` method as a fallback in case `GetPropertyValue()` is empty.