Start off with attribute validation
In `TopicMappingService`, attribute validation (i.e., ensuring that the validation prescribed by attributes such as `[Required]`, `[MinValue()]`, &c. are satisfied) was done at the end, after the target object had been populated. That made sense, as that was the only point at which we could evaluate it. With the `ReverseTopicMappingService`, however, we have a completed Binding Model upfront, and want to validate it before making any modifications to the target topic. As such, this is moved from the bottom of the `SetPropertyAsync()` method to the top.