Flip the source/target collection settings
In the `TopicMappingService`, there's an existing collection on the source `Topic` that needs to be retrieved, and then there's a collection definition on the target view model that may need to be created. In `ReverseTopicMappingService`, we know both exist, and there's only one location to pull from the source (the property we're currently on). As a result, the `GetSourceCollection()` is renamed to `GetTargetCollection()`, and largely operates as the original. Further, because we know we're dealing with Binding Models, we can assume that the source list on the Binding Model is of `ITopicBindingModel` to provide a touch more type safety. The target logic will need to be improved a bit further in a future commit, however, as not all collections should be updated directly (e.g., relationships should use `SetRelationship()`).