Initial check for scalar values
In the `TopicMappingService`, it was assumed that a value was a scalar if the target view model property was settable by the `TypeMemberInfoCollection` (i.e., is a `string`, `bool`, `int`, or `DateTime` object). For a first order approximation we can flip this by looking at the _target_ for the `ReverseTopicMappingService` and asking the same question. If so, we have some confidence we can map an `topic.Attributes.GetValue()` to it. That said, we will want to do further validation in the future against the `ContentTypeDescriptor` to make sure that the attribute is real, and possibly even check the attribute type (e.g., that it's a `Checkbox` if it's targeting a `bool`).