Set value using attribute instead of property
Since we're now writing to the `Topic`, we needn't use reflection to try to set a property. Instead, we now write directly to the `topic.Attributes` collection using its `SetValue()` method. It's worth noting that the `AttributeValueCollection` will still use reflection to call a local property on `Topic`, if one exists with the same name as the attribute key. But that's all handled under the hood by `SetValue()` and isn't something we need to think about.