Remove support for `[FilterByAttribute()]`
The `[FilterByAttribute()]` attribute allows a collection to be filtered based on some attribute of the source collection, such as the `ContentType`. This is useful for creating a produce view targeting a specific user interface. For a Binding Model, however, we want to save all data, so the ability to filter the results isn't desirable. This is removed.