Remove checking of `IsDisabled`
Disabled topics should never be mapped to the user interface. But there isn't really an analogous concept of disabled Binding Models. We could, conceivably, check to see if the target `Topic` is disabled, but there isn't really a concern with _updating_ a disabled `Topic`—and, especially, if that's the instruction we're receiving from the caller. That level of validation, if necessary, should be provided by the controller, not presumed by the mapping service.