Updated list population logic for Binding Models
In addition to renaming `childTopic` to `childBindingModel` to better reflect the reverse scenario, `ReverseTopicMappingService` also bypasses the previous logic to allow a `Topic` to be added directly to a view model collection without mapping if it permitted `Topic`s. We never expect `Topic` on a Binding Model, so that logic can be removed; we'll always do mapping of any child collections, should we support them.