Throw exception when binding children
The `ReverseTopicMappingService` is not designed to handle children, with the possible exception of Nested Topics. Children introduce not only complexity, but also risk during the merging scenario. For instance, if a developer attempts to merge a Binding Model with an empty `Children` property with an existing topic, any children under that topic will be deleted—along with their descendents. This could be intentional. But it could very well be an accident (e.g., with the form failing to bind correctly to the collection). Since the expected use case of `IReverseTopicMappingService` won't typically make sense for handling children, it makes more sense not to support them. Having an exception here is really just meant for developers, so they get a warning if they ever try to map a Binding Model with this property. We could easily ignore this, but then the developer might be wondering why the `Children` aren't getting mapped. This provides clearly expectations.