Streamlined retrieval of target collection
In `TopicMappingService`, we needed to not just assess whether a relationship _existed_ but _also_ whether it should be _crawled_, which would vary based on how `MapAsync()` was called, and what attributes were on the target view model. That compexity doesn't exist for `ReverseTopicMappingService`, and so it can remove the complexity of the local `GetRelationship()` helper function, and replace it with a simple condition against `[RelationshipType]`.