Updated XML docs based on recent changes
Outside of the variable names, the documentation was largely retained from the original `TopicMappingService` source code. That didn't accurately describe the functionality. Given that, I went through and updated all of the XML docs and many of the inline comments to ensure they reflected the current functionality. In addition, added a couple of ### TODO's where appropriate.