Removed type checking
Type checking was necessary in `TopicMappingService` because we didn't know what type of object the `Topic` would be mapped to at compile time, and whether it would be compatible with the target list. With `ReverseTopicMappingService` we know the source object is compatible with the source list because it's already stored there. Hypothetically, the target list may be focused on a subtype of `Topic` so we _could_ add that check in, but that's a bit of an edge case since we rarely subclass `Topic` anymore—and, even if we did, we'd _want_ an exception to be thrown in that case, instead of just silently ignoring the mismatch. In the future, we may add in a check for this, but only so we can throw a more descriptve exception.