Set value on attribute, even if null
In `TopicMappingService`, we didn't want to set the value if it was null because that would overwrite the `[DefaultValue()]`, if it had previously been set. In the `ReverseMappingService` scenario, we still honor the `[DefaultValue()]`, but if it's not set, or the value is `null`, we _want_ to override the target attribute. (Obviously, we don't want to do this if the field is required—but that would have been handled by previous validation routines.)