Removed try/catch when mapping to collections
Previously, we had a `try/catch` statement when adding mapped objects to the target collection. This was primarily in case there was a key conflict with a `KeyedCollection` since that was difficult to detect given that everything was being evaluated as an `IList`. In those cases, we just swallow the error, since it's an edge scenario related to `[Flatten]` that isn't a priority to address. That said, we never want to potentially lose data when dealing with a binding model, and so such a scenario _should_ throw an exception—as should any other errors. As such, the defensive logic is removed. In addition, we no longer really need a helper function to aid with this since it's now ust one line, so that is also removed.