Moved `*MemberInfoCollection` to `Internal` namespace, and made public
The `(Type)MemberInfoCollection` classes are exclusively intended for use on the *OnTopic* product, not for customer consumption. They are not tested for broader or more comprehensive use cases. Moving them to `Internal` clearly denotes this. In addition, they have been moved to `public` so that they can be easily used by other assemblies that are part of the *OnTopic* library. This makes way for, for instance, moving the `Ignia.Topics.Mapping` namespace to its own assembly.