Moved `*Descriptor*` to `Metadata` namespace
The `ContentTypeDescriptor` and `AttributeDescriptor`, along with their associated strongly-typed collections, provide metadata for topics that is distinct from the topics themselves. To clarify this relationship, they are being moved to the new `Metadata` namespace. (They are, in fact, still topics with attributes, but they are treated differently since they primarily provide information about topics and their attributes.)