Established `Internal.Mapping` namespace
Some classes in the `Ignia.Topics.Mapping` are not intended for public consumption, and should never be directly exposed to the consumer, even if they are marked as public. For instance, the `PropertyConfiguration` class is effectively an internal helper for relaying parameters between methods in the mapping implementations. And the `RelationshipMap` is just a "decoder ring" that it uses to translate between the (public) `RelationshipType` and `Relationships` enums, which overlap in what they describe.