Integrated `ModelType` into property assessment
When assessing a property on a Binding Model, the attribute's `ModelType` is now factored in to provide context on the expectations for that property. So, for instance, if a property is a collection but the corresponding attribute is _not_ a `Relationship` or a `NestdTopic`, then the collection is not processed.