Split `SetCollectionValue()` into `SetRelationships()` and `SetNestedTopicsAsync()`
In the `TopicMappingService` it doesn't really matter if the collection is of relationships, children, nested topics, or incoming relationships. They are all populated the same way. This is not true of the reverse scenario. Relationships and Nested Topics are handled very differently. To account for this, the legacy `SetCollectionValue()` method is now split into two distinct methods, `SetAttributes()` and `SetNestedTopicsAsync()`.