Established default `RelatedTopicBindingModel`
Since relationships and references only need to describe one property—the `UniqueKey` of the target—there isn't a need for custom Binding Models. As such, the `Ignia.Topics.Models` namespace provides a default implementation of the `RelatedTopicBindingModel`. Of course, customers are free to override this if they have the need for additional properties in their application—those properties just won't be utilized by `ReverseTopicMappingService`.