Introduced defensive logic
Currently—shamefully!—the `Type` attribute usually ends in a file type (e.g., `.cshtml`). This should be removed at some point. In the meanwhile, it needs to be stripped when evaluating the model type. Added in conditional logic to handle the future scenario where it's removed from the database, which we should target for the upcoming MVC release of the editor.