Added fake `AttributeDescriptor`s to select `ContentTypeDescriptor`s
The `ReverseTopicRepository`, by design, attempts to validate against the `TopicRepository`'s `Root:Configuration:ContentTypes` definitions. The `FakeTopicRepository` that ships with `Ignia.Topics.Tests`, however, didn't contain actual `AttributeDescriptor` definitions for any content type definitions, thus resulting in the tests failing. To mitigate this, this update fills in the `AttributeDescriptor`s for `ContentTypeDescriptor` and for `AttributeDescriptor`. It also adds a local helper function to aid in these types of operations in the future.