Simplified property type detection
For the property type detection, we now _exclusively_ rely upon the `AttributeDescriptor.ModelType` instead of also looking for clues on the Binding Model (such as return type, or attributes). This is simpler, and less likely to swallow errors due to mismatches (as those will now be caught during the conversion process).