Fixed existing key lookup
Originally, we relied on a positive `Id` to determine if a Nested Topic already existed in the target `Topic`. That's generally a good rule in a production environment. But using `Key` is more reliable in a test environment, since we are working with topics that will not be saved to the database. `Key` is also more consistent in that the existing topic is previously looked up using the `Key`, so it avoids potential mismatches between `Id` and `Key` (highly unexpected, but still reasonable to rely on the same identifier across an operation).