Established separate `ReferenceTopicBindingModel`
Previously, the reference test case—a `DerivedTopic` property aliased to the `TopicId` attribute—was on the `BasicTopicBindingModel` class. Since `TopicId` is one of the canonical reference examples, and yet `BasicTopicBindingModel` is the base class for almost all of the other Binding Models in the unit test assembly, this limits the flexibiltiy of the tests (e.g., by allowing an invalid example to be setup, as done in `InvalidReferenceTypeTopicBindingModel`. To fix this, the `DerivedTopic` property is moved from `BasicTopicBindingModel` to the new `ReferenceTopicBindingModel`.