Updated nested topic validation logic
The nested topic validation logic previously relied on explicit setting via the `[Attribute(AttributeType.NestedTopic)]` to determine that the property is associated with a Nested Topic. This should not be assumed or relied upon. Instead, it has been updated to use the `AttributeDescriptor.ModelType` to determine this, which is a more accurate approach. In addition, the exception was cleaned up to be more precise.