Updated reference validation logic
Previously, when validating reference properties, the return type of the property was used to infer that the property was a topic reference. This has been updated to instead use the `AttributeDescriptor.ModelType`, which is more reliable. In addition, the exception message has been updated to be more precise, and consistent with the other exceptions.