Introduced `DisabledAttributeProperty` test
The `DisabledAttributeProperty` creates a (new) `DisabledAttributeTopicBindingModel` class and sets its `UnmappedAttribute` property. When mapping it, however, it doesn't anticipate an exception, since the `UnmappedAttribute` property is decorated with the `[DisableMapping]` attribute. Further, it validates that, in fact, the `ReverseTopicMappingService` disregarded the property, by ensuring that its corresponding attribute on the resulting `Topic` is `null`.