Ensure that missing relationships are deleted
When a Binding Model maps a relationships scope, any relationships not included in the source collection should be deleted—otherwise, the mapping will only ever _add_ relationships, but not allow the Binding Model to _remove_ **existing** relationships. This issue is resolved in this update—and the test case modified to account for this scenario.