Ensure `bool` are mapped correctly
To avoid ambiguity, `bool`s are serialized as "1" or "0" when converted to attributes. This is handled automatically if calling `topic.Attributes.SetBoolean()`. Since `ReverseTopicMappingService` is always going through `topic.Attributes.SetValue()`, however, it can't take advantage of this, and thus must do the conversion manually. Since the default serialization is to use "true" or "false", this needs to be corrected for. (Though, it should be noted, that any `bool` attributes mapped as properties on `Topic`, such as `Topic.IsHidden`, will automatically handle this conversion. Other attributes, however, will not be covered by this.)