Introduced `MenuViewComponentBase{T}`
The `MenuViewComponentBase{T}` replaces the legacy `LayoutControllerBase{T}`—and, in fact, works virtually identical to it, except that it's a `ViewComponent` instead of a `Controller`. This means that there's only one action—the `InvokeAsync()` method—and that it derives from `ViewComponent` not `Controller`. Outside of that, the code is virtually identical.