Swallow nested mapping errors
Previously, when a nested mapping operation failed to find a corresponding view model, it would return an `object`. The mapping operation would then likely reject it since `object` almost certainly can't be assigned from whatever type the property uses. `TopicMappingService` was recently updated to instead throw an exception instead of falling back to `object`. This makes sense for entry-level calls. But for nested calls, it breaks legacy functionality—which is just to ignore these as mapping mismatches. To maintain that functionality, the error is swallowed. (If it weren't, and an object were returned, it would have been silently rejected anyway on the next line.)