Accounted for the `MenuViewComponent` in the documentation
Provided a basic update to account for the new `MenuViewComponent`. Since view components are so similar to controllers, they're currently merged into the same section. If we end up with additional view components in the future, we may want to separate this out into its own section. In addition, I renamed _Controller Activator_ to _Composition Root_ to account for the fact that we'll now need _both_ a controller activator _and_ a view component activator, at least if the `MenuViewComponent` (or any other view component) is going to be used. The **Gist** sample will need to be updated to account for this later.