Established `SetCompatibleProperty()` and tests
The `SetCompatibleProperty()` is a fallback method which will look for a 1:1 mapping of properties between a source and target object. It makes no attempt to convert the data (e.g., from a string to a strongly-typed object, as it does with attributes). This is a fallback to provide support for e.g., enums or other objects that may be added to a `Topic` (or its derivatives) but aren't supported by any of the converters.