Introduced `TopicDataModel`
The `TopicDataModel` is a super lightweight data container for representing a topic entity without any of the overhead or complexity of validation or referential integrity. It's intended to aid in binding, conversion, and serialization related tasks where a lightweight intermediary will make it easier to provide a general mapping service for handling a variety of different data sources. E.g., JSON or form data can each be mapped to `TopicDataModel`, which in turn can e mapped to a `Topic`.