Added `ReadJson` support for `RelatedTopicCollection` deserialization
While it's not an expected use case for the out-of-the-box serialization requirements, it is nice-to-have that the JSON conversion support deserialization as well as serialization for `Topic`. To facilitate this, added the `ReadJson()` implementation to the `RelatedTopicCollectionJsonConverter` class. Note that this _necessitates_ that the `JsonConverter` be applied via `JsonConvert.SerializeObject()`, or the converter won't be applied; this is a frustrating limitation, but necessary since a dependency (an `ITopicRepository`) is necessary for the `ReadJson` case.