Enable nullable annotation context
One of the major benefits (to us) for C# 8.0 is the ability to default reference types to non-nullable. Unfortunately, violations aren't enforced by the compiler, and instead only generate an error, and thus guard clauses are still required for public interfaces. Nevertheless, the ability to better indicate the design intent of classes should help reduce inadvertent `NullReferenceException` errors.