Remove null initialization for dependency fields
Previously, private backing fields for (readonly) dependencies were initialized as null, as a matter of consistency. This is unnecessary because they are, by design, always assigned to via the constructor (following the constructor injection pattern). Since these are never actually intended to be null, these should instead be left unassigned until the constructor assigns them their actual value. This aids with C# 8.0's nullable annotation context by helping establish that the backing field is not intended to be nullable.