Mark nullable types
As part of C# 8.0's nullable annotation contexts, any reference types that potentially expect a null value should be annotated with a `?` to make them nullable; without this a warning will be thrown. In many cases, values expect a potentially null value (e.g., when the default return value is null, or a parameter is optional). In other cases, a null value may not be desired—but is validated as part of the method, instead of prevalidating it prior to every method call. This commit provides an initial assessment of these changes in member types, variable declarations, variable assignments, and method parameters. Additional adustments will almost certainly be required later as further cases are identified—and, potentially, existing cases are reevaluated.