Additional null initialization for dependency
As per commit #96f34b6, dependencies should not be initialized to `null` since they are not intended to be nullable; instead, they should be declared, and then initialized to the correct instance value as part of the constructor. The `CachedHierarchicalTopicMappingService<T>` was missed as part of the original coverage of this.