Explicitly state fallback for `GetValue()`
Some `AttributeValueCollection.GetValue()` calls specify a default of blank (""). The compiler doesn't recognize these as a non-nullable fallback, however. As a (redundant) shorthand to prevent the nullable annotation context for mistaking these as null outputs, these are setup with an explicit fallback. Ideally, we'll find a better solution to these cases at a later date; for now, however, they're an economical way to inform the compiler that no null will be returned.