Remove null initialization for inline variables
Some local variables and private fields are never intended to be nullable and are only set to `null` as a default (in the case of auto-implemented properties) or as a placeholder (until they are later initialized with a correct value). This sends mixed messages to the compiler, implying these are nullable variables. To prevent confusion (and avoid an unnecessary initialization on immutable structs) these values are either being left uninitialized after declaration, or explicitly initialized to a default value.