Replace implicit null checks with explicit null checks
In place of implicit safe navigation operator ("?."), explicitly check for null values for variables that are never intended to be set to `null`. The safe navigation operator provides a convenient and economical shorthand, but it also implies nullability of the assignment unless a default is provided. Explicitly guarding against this (via a condition, or an actual guard clause) helps clarify the intent here, while avoiding downstream consequences of unexpected nulls.